refactor: standardize module naming conventions
Standardize naming conventions across all modules for consistency and clarity.

**Changes:**

**Posts Module:**
- Renamed `post_bindings.dart` → `posts_bindings.dart`
- Renamed class `PostBindings` → `PostsBindings`
- Updated all imports in binding.dart, posts.dart

**Connections Module:**
- Renamed `connection_bindings.dart` → `connections_bindings.dart`
- Renamed class `ConnectionBindings` → `ConnectionsBindings`
- Updated imports in main.dart and connection.dart barrel export

**Locale Module:**
- Renamed `data/model/` → `data/models/` (standardize to plural)
- Updated all imports in:
  - locale.dart
  - localization_view_model.dart
  - language_dialog.dart
  - All test files (language_model_test.dart, language_picker_dialog_test.dart, localization_view_model_test.dart)

**Documentation:**
- Updated coding_guidelines.md with explicit naming conventions:
  - Section 2.1: Module Structure with emphasis on `controllers/` (not `viewmodels/`)
  - Section 2.2: Module Naming Conventions
    - Module folders: plural (posts/, connections/)
    - Bindings files must match module name (posts_bindings.dart)
    - Bindings classes must match module name (PostsBindings)
    - Model folders: always plural `models/`
    - Controllers folder: always `controllers/` (GetX convention)
  - Added complete example structure
  - Section 2.3: Infrastructure

**Rationale:**
- Bindings file names now match module names for consistency
- `models/` folder is now plural everywhere (was inconsistent)
- `controllers/` directory clarified as GetX convention (not `viewmodels/`)
- Plural module names for collections (posts, connections)

**Verification:**
- All analyzer errors fixed
- All tests updated with correct imports
- No breaking changes to public APIs
