+++ lib/src/modules/connections/connection.dart
@@ -3,3 +3,4 @@ export 'views/connection_overlay.dart';
 export 'views/connectivity_checker_splash.dart';
 export 'controllers/connection_view_model.dart';
 export 'actions/connection_actions.dart';
+export 'connections_bindings.dart';
