+++ lib/src/modules/connections/connections_bindings.dart
@@ -1,13 +1,13 @@
 import 'package:get/get.dart';
 import 'controllers/connection_view_model.dart';
 
-/// ConnectionBindings
+/// ConnectionsBindings
 ///
 /// Ensures ConnectionViewModel is available before any widgets that depend
 /// on it (e.g., ConnectionOverlay) are built. Use this binding early in
 /// app startup (e.g., in main() before runApp) since ConnectionOverlay is
 /// wrapped outside GetMaterialApp where initialBinding would not yet run.
-class ConnectionBindings extends Bindings {
+class ConnectionsBindings extends Bindings {
   @override
   void dependencies() {
     if (!Get.isRegistered<ConnectionViewModel>()) {
