+++ lib/src/modules/locale/locale.dart
@@ -4,7 +4,7 @@
 // across the application.
 
 export 'controllers/localization_view_model.dart';
-export 'data/model/language_model.dart';
+export 'data/models/language_model.dart';
 export 'data/services/localization_service.dart';
 export 'views/language_dialog.dart';
 export 'views/language_drawer_item.dart';
