+++ test/modules/locale/language_model_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:getx_starter/src/modules/locale/data/model/language_model.dart';
+import 'package:getx_starter/src/modules/locale/data/models/language_model.dart';
 
 void main() {
   group('LanguageModel', () {
