+++ test/modules/locale/language_picker_dialog_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:get/get.dart';
 import 'package:getx_starter/src/modules/locale/controllers/localization_view_model.dart';
-import 'package:getx_starter/src/modules/locale/data/model/language_model.dart';
+import 'package:getx_starter/src/modules/locale/data/models/language_model.dart';
 import 'package:getx_starter/src/modules/locale/views/language_dialog.dart';
 import 'package:getx_starter/src/modules/locale/data/services/localization_service.dart';
